# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /collaboration-groups/{id}:
    patch:
      operationId: updateCollaborationGroupName
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      - in: query
        name: groupName
        type: string
        required: true
      responses:
        200:
          description: Updated the collaboration group name successfully
        401:
          description: Invalid token. No user was found for the provided token
        404:
          description: Specified CollaborationGroup does not exist
    delete:
      operationId: deleteCollaborationGroup
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Deleted the collaboration group successfully
        401:
          description: Invalid token. No user was found for the provided token
    get:
      operationId: getCollaborationGroup
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Retrieved the collaboration group successfully
          schema:
            $ref: '#/definitions/CollaborationGroup'
        401:
          description: Invalid token. No user was found for the provided token
        404:
          description: The specified CollaborationGroup does not exist
  /collaboration-groups/{id}/archive:
      post:
        operationId: archiveCollaborationGroup
        parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - in: path
          name: id
          type: string
          description: Identifier of the collaboration group to be archived
          required: true
        responses:
          200:
            description: Archived the collaboration group successfully
            schema:
              $ref: '#/definitions/CollaborationGroup'
          401:
            description: Invalid token. No user was found for the provided token
          404:
            description: The collaboration group does not exist for the given id
  /collaboration-groups/{id}/restore:
      post:
        operationId: restoreCollaborationGroup
        parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - in: path
          name: id
          type: string
          description: Identifier of the collaboration group to be restored
          required: true
        responses:
          200:
            description: Restored the collaboration group successfully
            schema:
              $ref: '#/definitions/CollaborationGroup'
          401:
            description: Invalid token. No user was found for the provided token
          404:
            description: The collaboration group does not exist for the given id
  /collaboration-groups/merge:
    get:
      operationId: mergeCollaborationGroups
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - in: query
          name: bcid
          type: string
          description: Identifier of the base collaboration group
          required: true
        - in: query
          name: cgids
          type: array
          description: List of collaboration group id's to be merged.
          required: true
      responses:
        200:
          description: Merged the collaboration group successfully
          schema:
            $ref: '#/definitions/CollaborationGroup'
        401:
          description: Invalid token. No user was found for the provided token
        400:
          description: Invalid Clause content
  /collaboration-groups/all-finished:
      get:
        operationId: checkAllCollaborationsFinished
        parameters:
          - name: Authorization
            in: header
            type: string
            required: true
          - in: query
            name: partyId
            type: string
            description: Identifier of the party whose collaborations are checked
            required: true
          - in: query
            name: collaborationRole
            type: string
            description: Role of the party in the collaboration. Possible values are SELLER and BUYER
            default: SELLER
            required: true
        responses:
          200:
            schema:
              type: string
          401:
            description: Invalid token. No user was found for the provided token
  /collaboration-groups:
    get:
      operationId: getCollaborationGroups
      parameters:
      - in: query
        name: partyId
        type: string
        description: Identifier of the party
      - in: query
        name: relatedProducts
        type: array
        description: Related products
      - in: query
        name: relatedProductCategories
        type: array
        description: Related product categories
      - in: query
        name: tradingPartnerIDs
        type: array
        description: Identifier of the corresponsing trading partner ID
      - in: query
        name: offset
        type: integer
        description: Offset of the first result among the complete result set satisfying the given criteria
        default: 0
      - in: query
        name: limit
        type: integer
        description: Number of results to be included in the result set
        default: 10
      - in: query
        name: archived
        type: boolean
        default: false
      - in: query
        name: status
        required: false
        type: array
        items:
          type: string
          enum:
          - STARTED
          - CANCELLED
          - COMPLETED
      - in: query
        name: collaborationRole
        type: string
        required: false
      - in: query
        name: isProject
        type: boolean
        default: false
      - name: Authorization
        in: header
        type: string
        required: true
      produces:
      - application/json
      responses:
        200:
          description: Retrieved the groups successfully
          schema:
            $ref: '#/definitions/CollaborationGroupResponse'
...
definitions:
...
  ProcessInstanceGroup:
    title: Process Instance Group Object
    type: object
    properties:
      ID:
        type: string
      name:
        type: string
      partyID:
        type: string
      dataChannelId:
        type: string
      processInstanceIDs:
        type: array
        items:
          type: string
      archived:
        type: boolean
      collaborationRole:
        type: string
      firstActivityTime:
        type: string
      lastActivityTime:
        type: string
      precedingProcessInstanceGroup:
        $ref: '#/definitions/ProcessInstanceGroup'
      status:
        type: string
        enum:
        - INPROGRESS
        - CANCELLED
  CollaborationGroup:
    title: Collaboration Group
    type: object
    properties:
      ID:
        type: string
      name:
        type: string
      archived:
        type: boolean
      isArchiveable:
        type: boolean
      status:
        type: string
        enum:
        - INPROGRESS
      isProject:
        type: boolean
      associatedProcessInstanceGroups:
        type: array
        items:
          $ref: '#/definitions/ProcessInstanceGroup'
  CollaborationGroupResponse:
    title: Process Instance Group Response
    type: object
...